+++ src/to_py.rs
@@ -1,9 +1,10 @@
+use crate::PyCoordNum;
 use geo_types::{
     Coordinate, Geometry, GeometryCollection, Line, LineString, MultiLineString, MultiPoint,
     MultiPolygon, Point, Polygon,
 };
 use pyo3::types::{PyDict, PyString, PyTuple};
-use pyo3::{intern, IntoPy, PyObject, PyResult, Python, ToPyObject};
+use pyo3::{intern, PyObject, PyResult, Python, ToPyObject};
 use std::iter::once;
 
 /// Convert `self` to a Python dictionary reflecting the structure of a `__geo_interface__` python dict.
@@ -11,7 +12,10 @@ pub trait AsGeoInterfacePyDict {
     fn as_geointerface_pydict<'py>(&self, py: Python<'py>) -> PyResult<&'py PyDict>;
 }
 
-impl AsGeoInterfacePyDict for Geometry<f64> {
+impl<T> AsGeoInterfacePyDict for Geometry<T>
+where
+    T: PyCoordNum,
+{
     fn as_geointerface_pydict<'py>(&self, py: Python<'py>) -> PyResult<&'py PyDict> {
         match self {
             Geometry::Point(g) => g.as_geointerface_pydict(py),
@@ -28,13 +32,19 @@ impl AsGeoInterfacePyDict for Geometry<f64> {
     }
 }
 
-impl AsGeoInterfacePyDict for Point<f64> {
+impl<T> AsGeoInterfacePyDict for Point<T>
+where
+    T: PyCoordNum,
+{
     fn as_geointerface_pydict<'py>(&self, py: Python<'py>) -> PyResult<&'py PyDict> {
         make_geom_dict(py, intern!(py, "Point"), self.0.to_py(py))
     }
 }
 
-impl AsGeoInterfacePyDict for MultiPoint<f64> {
+impl<T> AsGeoInterfacePyDict for MultiPoint<T>
+where
+    T: PyCoordNum,
+{
     fn as_geointerface_pydict<'py>(&self, py: Python<'py>) -> PyResult<&'py PyDict> {
         make_geom_dict(
             py,
@@ -44,13 +54,19 @@ impl AsGeoInterfacePyDict for MultiPoint<f64> {
     }
 }
 
-impl AsGeoInterfacePyDict for LineString<f64> {
+impl<T> AsGeoInterfacePyDict for LineString<T>
+where
+    T: PyCoordNum,
+{
     fn as_geointerface_pydict<'py>(&self, py: Python<'py>) -> PyResult<&'py PyDict> {
         make_geom_dict(py, intern!(py, "LineString"), self.0.to_py(py))
     }
 }
 
-impl AsGeoInterfacePyDict for MultiLineString<f64> {
+impl<T> AsGeoInterfacePyDict for MultiLineString<T>
+where
+    T: PyCoordNum,
+{
     fn as_geointerface_pydict<'py>(&self, py: Python<'py>) -> PyResult<&'py PyDict> {
         make_geom_dict(
             py,
@@ -60,7 +76,10 @@ impl AsGeoInterfacePyDict for MultiLineString<f64> {
     }
 }
 
-impl AsGeoInterfacePyDict for Line<f64> {
+impl<T> AsGeoInterfacePyDict for Line<T>
+where
+    T: PyCoordNum,
+{
     fn as_geointerface_pydict<'py>(&self, py: Python<'py>) -> PyResult<&'py PyDict> {
         make_geom_dict(
             py,
@@ -70,14 +89,20 @@ impl AsGeoInterfacePyDict for Line<f64> {
     }
 }
 
-fn polygon_coordinates_to_pyobject(py: Python, polygon: &Polygon<f64>) -> PyObject {
+fn polygon_coordinates_to_pyobject<T>(py: Python, polygon: &Polygon<T>) -> PyObject
+where
+    T: PyCoordNum,
+{
     let linestring_objs: Vec<_> = once(polygon.exterior().0.to_py(py))
         .chain(polygon.interiors().iter().map(|ls| ls.0.to_py(py)))
         .collect();
     PyTuple::new(py, linestring_objs).to_object(py)
 }
 
-impl AsGeoInterfacePyDict for Polygon<f64> {
+impl<T> AsGeoInterfacePyDict for Polygon<T>
+where
+    T: PyCoordNum,
+{
     fn as_geointerface_pydict<'py>(&self, py: Python<'py>) -> PyResult<&'py PyDict> {
         make_geom_dict(
             py,
@@ -87,7 +112,10 @@ impl AsGeoInterfacePyDict for Polygon<f64> {
     }
 }
 
-impl AsGeoInterfacePyDict for MultiPolygon<f64> {
+impl<T> AsGeoInterfacePyDict for MultiPolygon<T>
+where
+    T: PyCoordNum,
+{
     fn as_geointerface_pydict<'py>(&self, py: Python<'py>) -> PyResult<&'py PyDict> {
         make_geom_dict(
             py,
@@ -114,7 +142,10 @@ fn make_geom_dict<'py>(
     Ok(dict)
 }
 
-impl AsGeoInterfacePyDict for GeometryCollection<f64> {
+impl<T> AsGeoInterfacePyDict for GeometryCollection<T>
+where
+    T: PyCoordNum,
+{
     fn as_geointerface_pydict<'py>(&self, py: Python<'py>) -> PyResult<&'py PyDict> {
         let dict = PyDict::new(py);
         dict.set_item(intern!(py, "type"), intern!(py, "GeometryCollection"))?;
@@ -136,13 +167,19 @@ trait ToPy {
     fn to_py(&self, py: Python) -> PyObject;
 }
 
-impl ToPy for Coordinate<f64> {
+impl<T> ToPy for Coordinate<T>
+where
+    T: PyCoordNum,
+{
     fn to_py(&self, py: Python) -> PyObject {
         PyTuple::new(py, &[self.x.into_py(py), self.y.into_py(py)]).to_object(py)
     }
 }
 
-impl ToPy for [Coordinate<f64>] {
+impl<T> ToPy for [Coordinate<T>]
+where
+    T: PyCoordNum,
+{
     fn to_py(&self, py: Python) -> PyObject {
         PyTuple::new(py, self.iter().map(|c| c.to_py(py))).to_object(py)
     }