+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## Unreleased
+
+* Generic `GeoInterface` implementations for `f64`, `f32`, `u8`, `u16`, `u32`, `u64`, `i8`, `i16`, `i32` and `i64`.
+
 ## 0.1.1
 
 * Support for `Py_LIMTED`-API by avoiding usage of `PyTuple::as_slice()`